src/web_app/requirements.txt
flask>=3.0.0
markdown>=3.5.0
python-dotenv>=1.0.0